+++ ChangeLog.md
@@ -9,6 +9,9 @@
  - Travis Continuous integration will check that DGtalTools still compiles with
    changes in new pull-requests. (David Coeurjolly,
    [#1133](https://github.com/DGtal-team/DGtal/pull/1133))
+ - Add cmake configuration file NeighborhoodTablesConfig to decompress and install look up tables. (Pablo Hernandez-Cerdan, [#1155](https://github.com/DGtal-team/DGtal/pull/1155))
+
+
 
 - *Base Package*
  - Alias and ConstAlias now raise compilation error when using invalid
@@ -74,6 +77,7 @@
   - Fix interior/exterior fill methods of topology/helpers/Surfaces class which
     was wrong on 3d and on closed Khalimsky space.
     (Bertrand Kerautret, [#1156](https://github.com/DGtal-team/DGtal/pull/1156))
+  - Add pre-calculated look up tables to speed up Object::isSimple calculations. (Pablo Hernandez-Cerdan, [#1155](https://github.com/DGtal-team/DGtal/pull/1155))
 
 - *Shape Package*
   - Fix a tubular mesh construction problem (missing faces) which appears